# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "KeepNote API",
    "description": ""
  },
  "host": "localhost:8000",
  "basePath": "/",
  "tags": [
    {
      "name": "Notes",
      "description": "Endpoints"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/notes": {
      "post": {
        "tags": [
          "Notes"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "any"
                },
                "body": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      },
      "get": {
        "tags": [
          "Notes"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/notes/{id}": {
      "get": {
        "tags": [
          "Notes"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      },
      "delete": {
        "tags": [
          "Notes"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      },
      "put": {
        "tags": [
          "Notes"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    }
  },
  "definitions": {
    "Notes": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "Buy Grocery"
        },
        "body": {
          "type": "string",
          "example": "Mon, Wed, Fri schedule to buy groceries"
        }
      }
    },
    "AddNote": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "Buy Grocery"
        },
        "body": {
          "type": "string",
          "example": "Mon, Wed, Fri schedule to buy groceries"
        }
      },
      "required": [
        "title",
        "body"
      ]
    }
  }
}